change pet owner window
