+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.LogginApp"
+        android:theme="@style/Theme.AppCompat.Light"
         tools:targetApi="31">
         <activity
             android:name=".MakeAppointment"
@@ -37,22 +37,26 @@
         <activity
             android:name=".UpdatePetDetails"
             android:exported="false" />
+<!--        <activity android:name=".UpdatePetDetails2"-->
+<!--            android:exported="false" />-->
         <activity
             android:name=".VetWindow"
-            android:exported="false" />
+            android:exported="false"
+            android:theme="@style/Theme.AppCompat.Light"/>
         <activity
             android:name=".VeterinarianSignUp"
             android:exported="false" />
         <activity
             android:name=".PetOwnerWindow"
             android:exported="false"
-            android:theme="@style/Theme.LogginApp" />
+            android:theme="@style/Theme.AppCompat.Light"/>
         <activity
             android:name=".OwnerSignUp"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/Theme.AppCompat.Light">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -61,7 +65,8 @@
         </activity>
         <activity
             android:name=".SignUp"
-            android:exported="false" />
+            android:exported="false"
+            android:theme="@style/Theme.AppCompat.Light"/>
     </application>
 
 </manifest>
