+++ app/src/main/res/values/themes.xml
@@ -50,5 +50,6 @@
 
 
     </style>
+    <style name="Theme.LogginApp" parent="Base.Theme.LogginApp" />
 
 </resources>
